Factored UdapDynamicClientRegistrationDocument fom Udap.Server to Udap.Common
Moved system style tests out of UdapServer.Tests
