+++ _tests/WebApi.Tests/FhirLabsApi/UdapControllerCommunityTest.cs
@@ -68,7 +68,7 @@ public class ApiForCommunityTestFixture : WebApplicationFactory<program>
 
     protected override IHost CreateHost(IHostBuilder builder)
     {
-        builder.UseEnvironment("Development");
+        builder.UseEnvironment("development");
         builder.ConfigureLogging(logging =>
         {
             logging.ClearProviders();
