+++ _tests/WebApi.Tests/WeatherApi/UdapControllerCommunityTest.cs
@@ -68,7 +68,7 @@ public class ApiForCommunityTestFixture : WebApplicationFactory<program>
 
     protected override IHost CreateHost(IHostBuilder builder)
     {
-        builder.UseEnvironment("Development");
+        builder.UseEnvironment("development");
         builder.ConfigureLogging(logging =>
         {
             logging.ClearProviders();
