+++ examples/FhirLabsApi/appsettings.release.json
@@ -15,9 +15,9 @@
         "SignedMetadataConfig": {
           "Issuer": "https://fhirlabs.net:7016/fhir/r4",
           "Subject": "https://fhirlabs.net:7016/fhir/r4",
-          "AuthorizationEndPoint": "http://localhost:5002/connect/authorize",
-          "TokenEndpoint": "http://localhost:5002/connect/token",
-          "RegistrationEndpoint": "http://localhost:5002/connect/register"
+          "AuthorizationEndPoint": "http://securedcontrols.net:5002/connect/authorize",
+          "TokenEndpoint": "http://securedcontrols.net:5002/connect/token",
+          "RegistrationEndpoint": "http://securedcontrols.net:5002/connect/register"
         }
       },
       {
@@ -25,9 +25,9 @@
         "SignedMetadataConfig": {
           "Issuer": "https://stage.healthtogo.me:8181",
           "Subject": "https://stage.healthtogo.me:8181",
-          "AuthorizationEndPoint": "https://localhost:5002/connect/authorize",
-          "TokenEndpoint": "https://localhost:5002/connect/token",
-          "RegistrationEndpoint": "https://localhost:5002/connect/register"
+          "AuthorizationEndPoint": "https://securedcontrols.net:5002/connect/authorize",
+          "TokenEndpoint": "https://securedcontrols.net:5002/connect/token",
+          "RegistrationEndpoint": "https://securedcontrols.net:5002/connect/register"
         }
       },
       {
